refactor(accounting): move accounting types to a new file
